Implement brick tracking
The last brick added to the scene is now accessible via lastBrick; a debug log spits out this brick’s coordinates as it moves. This can be used to determine when to add more bricks.

Bricks are organized by category instead of name, because it was hard to use wildcards in brick names.