Refactor brick add method
Adds a brickContainer method to create the brick container separately so that we can append to it. I was creating a new brickContainer (I think) every time I added a group of bricks, and as soon as the new one was created I lost my reference to the previous bricks, so I couldn’t remove them.